requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.24.0
streamlit>=1.28.0
easyocr>=1.7.0
torch>=2.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0
paddleocr>=2.8.0
